v0.8
Finalização de todas as classes

+++ Res1_agenda_estudantil/src/Classes/Estudante.java
@@ -24,7 +24,7 @@ public class Estudante {
             return Curso;
         }
 
-        public float get_Periodo() {
+        public int get_Periodo() {
             return Periodo;
         }
